+++ aok/__init__.py
@@ -1,3 +1,3 @@
 from _aok_version import version as __version__
 
-from .core import acquisition, kd_utils, main
+from .core import datarequest, kd_utils, main
